Reorganize worklist exceptions into the same module
