+++ robotools/evotools/__init__.py
@@ -1,3 +1,3 @@
-from robotools.evotools.exceptions import InvalidOperationError
 from robotools.evotools.types import Labwares, Tip, int_to_tip
 from robotools.evotools.worklist import EvoWorklist, Worklist
+from robotools.worklists.exceptions import InvalidOperationError
