--- robotools/evotools/exceptions.py
@@ -1,7 +1,0 @@
-"""Exceptions related to liquid handling with worklists."""
-
-__all__ = ("InvalidOperationError",)
-
-
-class InvalidOperationError(Exception):
-    pass
