+++ robotools/worklists/__init__.py
@@ -1,1 +1,2 @@
-from robotools.worklists.base import BaseWorklist, CompatibilityError
+from robotools.worklists.base import BaseWorklist
+from robotools.worklists.exceptions import CompatibilityError, InvalidOperationError
